Track move of disable_methods(), Initializer()
